Make InsertReshardings a preseg pass.
Similar to https://github.com/NVIDIA/Fuser/pull/2597.
