+++ csrc/multidevice/utils.h
@@ -86,11 +86,6 @@ int64_t requestedNumberOfDevices(Fusion*);
 void unshard(Fusion*);
 void unshard(TensorView*);
 
-// Runs through the fusion and inserts a resharding Set Op after
-// any resharding Expr that is not directly lowerable to a series of
-// communications
-void insertReshardings(Fusion* fusion);
-
 // This can only run after the insertResharding passes.
 // Assumes all resharding ops are either a set or reduction.
 // For each resharding operation that requires communication
