+++ tests/cpp/test_sharding.cpp
@@ -12,6 +12,7 @@
 #include <multidevice/executor.h>
 #include <multidevice/utils.h>
 #include <ops/all_ops.h>
+#include <preseg_passes/insert_reshardings.h>
 #include <preseg_passes/propagate_shardings.h>
 #include <tests/cpp/utils.h>
 #include <tests/cpp/validator.h>
@@ -104,7 +105,8 @@ TEST_F(ShardingTest, ShardedAllocationDomain) {
 
   preseg_passes::OptimizationPass<
       preseg_passes::PropagateShardingsPass>::runPass(&fusion);
-  insertReshardings(&fusion);
+  preseg_passes::OptimizationPass<
+      preseg_passes::InsertReshardingsPass>::runPass(&fusion);
   insertShardedAxisReordering(&fusion);
   setShardedAllocationDomain(&fusion);
   for (auto expr : fusion.exprs()) {
